# Add new `Performance/MapMethodChain` cop