--- a/changelog/new_add_new_performance_map_method_chain_cop.md
+++ b/changelog/new_add_new_performance_map_method_chain_cop.md
@@ -1,0 +1,1 @@
+* [#364](https://github.com/rubocop/rubocop-performance/pull/364): Add new `Performance/MapMethodChain` cop. ([@koic][])